Update WeightedStrings.java
In Hacker Land every character has a weight. The weight of an , In Hacker Land every character has a weight. The weight of an English uppercase alphabet A-Z is given below.asy -A = 1B = 2*A + AC = 3*B +  In Hackerland every character has a weight. The weight of an English uppercase alphabet A-Z is given below : A = 1. B = 2*A + A. C = 3*B + B . D = 4*C + C …. Z = 26*Y + Y. The weight made up of these characters is the summation of weights of each character. Given a total string weight, determine shortest string of given weight